Only include predictions if a departure prediction is present (#159)
* Only include predictions if a departure prediction is present

If a prediction has an arrival value only, and no departure, this means the train will not be continuing in this direction through this station. For example, when a shuttle-bus diversion is happening, a station that isn't normally the last stop becomes the last stop which is indicated by an arrival prediction with a nil departure prediction. In these cases, we do not want to display the prediction on any signs so this commit filters out such predictions.

* Name test stop in a helpful way

Instead of a meaningless stop_id like "10" use a stop ID that explains the purpose of the stop in the test.