Treat "stopped at station" as a regular BRD
There were two different types of situations that both resulted in "BRD" being displayed - one was a "prediction message" with :boarding as the number of minutes, and the other a "stopped train" message with "zero stops away."

This caused a problem by short-circuiting the logic that differentiates between terminals and non-terminals for the purposes of determining when to put "BRD" on the signs. This PR fixes that issue in (I believe) the simplest way possible.