fix: crash related to DateTimes
Timex.parse!/2 can return a time in the "Etc/UTC-4" zone, which is
invalid according to IANA and not present in the Timezone DB we were
using, Tzdata.Database. Any DateTimes returned by that function, can
then not be used with any stdlib functions that are timezone aware,
like DateTime.add/2.