fix: support separate alert status for top/bottom
different alert status per top/bottom config are needed to support
showing an alert status based message on a single source config. this is
most common when top and bottom source configs are for different
lines/routes.

previously the alert_status for a sign was reduced down to a single
value which could cause weird behavior when an alert was only effecting
one of the two source configs.