feat: signs can have a "default mode"
Previously, if a sign wasn't configured in SignsUI, it would always be
in the `off` mode. We'd like to be able to configure our new lab signs
for testing without the complexity of supporting this unique "kind" of
sign in SignsUI as well as this project.

This adds an optional `default_mode` field to sign configuration, which
can be set to any of the "modes" supported by SignsUI. For now, this is
only supported by `Realtime` (a.k.a. subway) signs and not `Bus` signs,
but this would be easy enough to extend later if needed.

Due to now needing a sign's default mode as well as its ID whenever we
want to determine its "current mode", this includes a refactor of the
previously-added `current_config` logging so the value is determined
when a message is enqueued, instead of when it is later processed.