fix: expand handling of alert status on signs with top/bottom configs (#797)
* fix: support separate alert status for top/bottom

different alert status per top/bottom config are needed to support
showing an alert status based message on a single source config. this is
most common when top and bottom source configs are for different
lines/routes.

previously the alert_status for a sign was reduced down to a single
value which could cause weird behavior when an alert was only effecting
one of the two source configs.

* chore: refactor get_paging_headway_or_alert_messages

Moves multiple function head pattern matching with two `case`
expressions to reduce the amount of duplilcation in the function bodies.

* fix: handle split alert statuses for signs in headway mode

@PaulJKim pointed out that mezzanine signs when set to `:headway` mode
check alert status but that code hadn't been updated to handle when
alert statuses are different between `top` and `bottom` source
configurations.

This change intercepts split alert status values passed to
`get_alert_messages/2` and gets the highest priority alert status of the
two using `Engine.Alerts.higher_priority_status/2` before continuing on
to the existing implementation.